+++ NEWS.md
@@ -1,4 +1,5 @@
 # themask (development version)
+* Store masked MSD dataset as a Github Release [2023-10-02] 
 * Utilize `gagglr` update check [2023-10-02]
 * Add `pkgdown` site [2023-10-02]
 * Add basic tests [2023-10-02]